IAbility WIP known 2504 error
I'm loading this up to branch because I need help with it and I'm hoping to get someone to review and help me diagnose what is going wrong.
