+++ ConsoleApplication1/gameObjectClasses/Buff.h
@@ -1,6 +1,7 @@
 #pragma once
 #include "GameObject.h"
 #include "Mod.h"
+#include "../sys/IAbilityDirector.h"
 
 
 class Buff :
