+++ ConsoleApplication1/interfaceClasses/IAbilityClasses/DecrementFactory.cpp
@@ -1,0 +1,9 @@
+#pragma once
+
+#include "DecrementFactory.h"
+
+
+std::shared_ptr<IAbility> DecrementFactory::makeIAbility(std::shared_ptr<GameObject> owner)
+{
+    return std::make_shared<Decrement>(owner);
+}
